# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP-сервер на Go с автоматической генерацией Swagger при старте",
        "title": "Configurator",
        "contact": {
            "name": "Lorgar Avrelian",
            "url": "https://github.com/Lorgar-Avrelian",
            "email": "victor-14-244@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://apache.org"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
        "/api/v1/components": {
            "get": {
                "description": "Возвращает список всех составных частей устройств вместе со всеми собственными и унаследованными параметрами",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "1. Модельный каталог: Компоненты"
                ],
                "summary": "Получить все компоненты вместе с параметрами",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurator_internal_model.Component"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новую запись составной части устройства. Если base_component равен null, создается базовый компонент.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "1. Модельный каталог: Компоненты"
                ],
                "summary": "Создать компонент",
                "parameters": [
                    {
                        "description": "Данные компонента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_dto.ComponentCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_model.Component"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/components/search": {
            "get": {
                "description": "Ищет компоненты, у которых строка совпадает с title, name_en, name_ru, description_en или description_ru",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "1. Модельный каталог: Компоненты"
                ],
                "summary": "Поиск компонентов по подстроке",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Строка поиска",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurator_internal_model.Component"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/components/{id}": {
            "get": {
                "description": "Возвращает компонент и иерархически объединенный список его параметров из БД",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "1. Модельный каталог: Компоненты"
                ],
                "summary": "Получить компонент по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Компонента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_model.Component"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "1. Модельный каталог: Компоненты"
                ],
                "summary": "Обновить компонент",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Компонента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_dto.ComponentUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_model.Component"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет компонент по ID, а также автоматически стирает всю иерархию дочерних элементов",
                "tags": [
                    "1. Модельный каталог: Компоненты"
                ],
                "summary": "Удалить компонент и всех его наследников",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Компонента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "configurator_internal_dto.ComponentCreate": {
            "type": "object",
            "required": [
                "access",
                "name_en",
                "name_ru",
                "title"
            ],
            "properties": {
                "access": {
                    "type": "string",
                    "example": "USER"
                },
                "base_component": {
                    "type": "integer",
                    "example": 1
                },
                "description_en": {
                    "type": "string",
                    "example": "Physical device - Device with dimensional characteristics"
                },
                "description_ru": {
                    "type": "string",
                    "example": "Физическое устройство - устройство, имеющее габаритные характеристики"
                },
                "name_en": {
                    "type": "string",
                    "example": "Physical device"
                },
                "name_ru": {
                    "type": "string",
                    "example": "Физическое устройство"
                },
                "title": {
                    "type": "string",
                    "example": "physical"
                }
            }
        },
        "configurator_internal_dto.ComponentUpdate": {
            "type": "object",
            "required": [
                "access",
                "name_en",
                "name_ru",
                "title"
            ],
            "properties": {
                "access": {
                    "type": "string",
                    "example": "USER"
                },
                "base_component": {
                    "type": "integer",
                    "example": 1
                },
                "description_en": {
                    "type": "string",
                    "example": "Updated main power controller unit"
                },
                "description_ru": {
                    "type": "string",
                    "example": "Обновленный главный контроллер питания"
                },
                "name_en": {
                    "type": "string",
                    "example": "power_module_v2"
                },
                "name_ru": {
                    "type": "string",
                    "example": "модуль_питания_в2"
                },
                "title": {
                    "type": "string",
                    "example": "Обновленный Модуль"
                }
            }
        },
...
        "configurator_internal_model.Component": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string",
                    "example": "USER"
                },
                "base_component": {
                    "type": "integer",
                    "example": 1
                },
                "description_en": {
                    "type": "string",
                    "example": "Main power controller unit"
                },
                "description_ru": {
                    "type": "string",
                    "example": "Главный контроллер питания"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "name_en": {
                    "type": "string",
                    "example": "power_module"
                },
                "name_ru": {
                    "type": "string",
                    "example": "модуль_питания"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/configurator_internal_model.Param"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Модуль Питания"
                }
            }
        },
...
        "configurator_internal_model.Param": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string",
                    "example": "ADMIN"
                },
                "description_en": {
                    "type": "string",
                    "example": "Maximum operational voltage"
                },
                "description_ru": {
                    "type": "string",
                    "example": "Максимальное рабочее напряжение"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name_en": {
                    "type": "string",
                    "example": "max_voltage"
                },
                "name_ru": {
                    "type": "string",
                    "example": "макс_напряжение"
                },
                "saved": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Макс. Напряжение"
                },
                "type": {
                    "type": "string",
                    "example": "INT"
                },
                "units_en": {
                    "type": "string",
                    "example": "V"
                },
                "units_ru": {
                    "type": "string",
                    "example": "В"
                },
                "value": {
                    "type": "string",
                    "example": "220"
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
...
}